requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.1
ansi2html==1.8.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==19.3.0
autopage==0.5.1
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cached-property==1.5.1
cachetools==5.3.0
captum==0.6.0
catboost==1.1.1
certifi==2019.11.28
certipy==0.1.3
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
clickhouse-driver==0.2.5
cliff==4.0.0
cmaes==0.8.2
cmake==3.26.3
cmd2==2.4.2
colorlog==6.7.0
comm==0.1.2
contourpy==1.0.6
convertdate==2.4.0
cryptography==38.0.4
cycler==0.11.0
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.16
deap==1.3.3
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
ephem==4.1.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.11.0
Flask==2.2.3
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
google-auth==2.17.2
google-auth-oauthlib==1.0.0
graphviz==0.20.1
greenlet==2.0.1
grpcio==1.53.0
hijri-converter==2.2.4
holidays==0.22
idna==2.8
imbalanced-learn==0.9.0
importlib-metadata==4.13.0
iotop==0.6
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==7.4.8
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-core==5.1.1
jupyter-dash==0.4.2
jupyter-events==0.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.1
jupyter-server==2.0.6
jupyter-server-terminals==0.4.3
jupyter-telemetry==0.1.0
jupyterhub==3.1.0
jupyterhub-idle-culler==1.2.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lightgbm==3.3.2
lightning-utilities==0.8.0
lit==16.0.1
llvmlite==0.40.0
LunarCalendar==0.0.9
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdstat==1.0.4
megacli==0.0.11
mistune==2.0.4
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==3.1
neuralprophet==0.5.4
notebook==6.5.2
notebook-shim==0.2.2
numba==0.57.0
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openpyxl==3.1.2
optuna==3.0.3
orjson==3.8.10
packaging==22.0
pamela==1.0.0
pandarallel==1.6.4
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
plotly==5.14.1
plotly-resampler==0.8.3.2
prettytable==3.4.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.22.1
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
PuLP==2.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycrypto==2.6.1
pycurl==7.45.2
Pygments==2.13.0
PyGObject==3.36.0
PyHive==0.6.5
pymacaroons==0.13.0
PyMeeus==0.5.12
PyNaCl==1.3.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.21.0
python-json-logger==2.0.4
pytorch-lightning==1.9.4
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.0
scipy==1.8.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
splunk-sdk==1.7.3
splunklib==1.0.0
SQLAlchemy==1.4.45
ssh-import-id==5.10
stack-data==0.6.2
statsmodels==0.14.0
stevedore==4.0.1
sympy==1.11.1
tenacity==8.1.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
texttable==1.6.2
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.1
torchmetrics==0.11.4
tornado==6.2
tqdm==4.65.0
trace-updater==0.0.9.1
traitlets==5.8.0
trino==0.320.0
triton==2.0.0
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==8001
ufw==0.36
uri-template==1.2.0
urllib3==1.25.8
varname==0.11.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==2.2.3
widgetsnbextension==4.0.5
xgboost==2.0.0
yarl==1.8.2
zipp==1.0.0
